+++ tests/ChefTest.php
@@ -109,4 +109,53 @@ class ChefTest extends AbstractChefTest
 
         self::assertArrayHasKey('callErrors', $topChefCalled);
     }
+
+    public function testErrorWithCatcherWithTopChefButErrorReportingIsStopped()
+    {
+        $topChef = $this->createMock(Chef::class);
+        $topChefCalled = [];
+        $topChef->expects(self::any())
+            ->method('__call')
+            ->willReturnCallback(function ($name) use ($topChef, &$topChefCalled) {
+                $topChefCalled[$name] = true;
+
+                return $topChef;
+            });
+
+        $chef = $this->buildChef($topChef);
+        $chef->read($this->createMock(RecipeInterface::class));
+
+        $called = false;
+        $bowl = $this->createMock(BowlInterface::class);
+        $bowl->expects(self::once())
+            ->method('execute')
+            ->willReturnCallback(function () use ($chef, &$called, $bowl) {
+                $called = true;
+                $chef->stopErrorReporting();
+                self::assertInstanceOf(
+                    ChefInterface::class,
+                    $chef->error(
+                        new \Exception('foo')
+                    )
+                );
+
+                return $bowl;
+            });
+
+        $errorBowl = $this->createMock(BowlInterface::class);
+        $errorBowl->expects(self::once())
+            ->method('execute')
+            ->willReturnSelf();
+
+        $chef->followSteps([$bowl], [$errorBowl]);
+
+        self::assertInstanceOf(
+            ChefInterface::class,
+            $chef->process(['foo'=>'bar'])
+        );
+
+        self::assertTrue($called);
+
+        self::assertArrayNotHasKey('callErrors', $topChefCalled);
+    }
 }